Added spring support

+++ src/main/java/com/robert/chatapp/Embeddable/UserGroupId.java
@@ -9,32 +9,32 @@ import java.util.Objects;
 public class UserGroupId implements Serializable {
 
     @Column(name = "user_id")
-    private int userId;
+    private Long userId;
 
     @Column(name = "group_id")
-    private int groupId;
+    private Long groupId;
 
     public UserGroupId() {
     }
 
-    public UserGroupId(int userId, int groupId) {
+    public UserGroupId(Long userId, Long groupId) {
         this.userId = userId;
         this.groupId = groupId;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public int getGroupId() {
+    public Long getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(int groupId) {
+    public void setGroupId(Long groupId) {
         this.groupId = groupId;
     }
